+++ subprojects/xcode-ide-kit/src/main/java/dev/nokee/xcode/PropertyListWriter.java
@@ -43,7 +43,7 @@ public interface PropertyListWriter extends Closeable {
 	void writeBoolean(boolean b);
 	void writeString(CharSequence s);
 	void writeInteger(long n);
-	void writeReal(float n); // TODO: Should be double
+	void writeReal(double n);
 
 	void flush();
 }
